Add support for stylePaths prop on TreeView
The stylePaths prop can be an Object map of JSONPath selectors to style
objects. The JSONPath selectors use the same syntax as the expandPaths
props.

If the current TreeNode matches one of the keys in 'stylePaths', the
styles for that key will be applied to the current node. For example:

stylePaths={{"$.id": { backgroundColor: "blue" } }}

will match the following object:

{
    id: "123", // This node will have a blue background
    otherProperty: "abc"
}