Add HTML_TAGNAME_TEXT_TRANSFORM to theme property
The hard-coded `textTransform` styling in the default theme is
problematic for XML documents that rely on uppercase tag names. In an
SVG document, for example, the `linearGradient` tag name should have a
capital 'G'. The current default style lowercases the entire tagname.

This diff adds `HTML_TAGNAME_TEXT_TRANSFORM` as a theme property so it
can be overridden by users, and maintains backwards compatibility by
keeping the default as 'lowercase'.

@xyc - does this look good to you? Curious if there was a reason you
added 'lowercase'? I figure it could be removed completely, but thought
we might as well keep it as an overridable property.